v4l: xilinx: dma: Fix INFINITE_LOOP warning
This patch fix INFINITE_LOOP type coverity warning.

warning:
Loop bound "*nplanes" has type "unsigned int", which is wider in
size or has a bigger upper bound than the type "u8" of loop counter
"i", the loop condition may always be true.

+++ drivers/media/platform/xilinx/xilinx-dma.c
@@ -323,7 +323,7 @@ xvip_dma_queue_setup(struct vb2_queue *vq,
 		     unsigned int sizes[], struct device *alloc_devs[])
 {
 	struct xvip_dma *dma = vb2_get_drv_priv(vq);
-	u8 i;
+	unsigned int i;
 	int sizeimage;
 
 	/* Multi planar case: Make sure the image size is large enough */
